requirement.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.5
Django==2.1.7
django-coreapi-client==1.1.1
django-cors-headers==2.4.0
django-filter==2.1.0
django-guardian==1.5.0
django-redis==4.10.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
msgpack==0.6.1
numpy==1.16.2
Pillow==5.4.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2018.9
redis==3.2.0
requests==2.21.0
selenium==3.141.0
six==1.12.0
torch==1.0.1
torchvision==0.2.2
tqdm==4.19.9
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
zope.interface==4.6.0
